requirements.txt
Flask==3.1.0
openai==1.61.1
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3